Hedgehog.Gen.QuickCheck

Description

Use QuickCheck generators and Arbitrary instances with Hedgehog.

Documentation

arbitrary :: (Arbitrary a, MonadGen m) => m a

Create a Hedgehog Gen from a QuickCheck Arbitrary instance.

The Arbitrary's shrink function is used to provide shrinking for the Hedgehog Gen.

quickcheck :: MonadGen m => Gen a -> m a

Create a Hedgehog Gen from a QuickCheck Gen.

By default the Gen created will not have any shrinking, you can use Gen.shrink if you have a shrink function which you would like to apply.